rename code snippets file
